requirements.txt
absl-py==0.11.0
astunparse==1.6.3
beautifulsoup4==4.9.3
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
dateparser==1.0.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
GoogleNews==1.5.2
grpcio==1.32.0
h5py==2.10.0
idna==2.10
jieba3k==0.35.1
joblib==1.0.0
Keras-Preprocessing==1.1.2
lxml==4.6.2
Markdown==3.3.3
newspaper3k==0.2.8
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scipy==1.4.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tinysegmenter==0.3
tldextract==3.1.0
tokenizers==0.9.4
tqdm==4.55.1
transformers==4.1.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
